backend/requirements.txt
Django>=5.1
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.1
psycopg2-binary>=2.9.9
python-decouple>=3.8
Pillow>=11.0.0
pytesseract>=0.3.10
pdf2image>=1.17.0
PyPDF2>=3.0.1
python-magic>=0.4.27
django-cors-headers>=4.4.0
gunicorn>=22.0.0
whitenoise>=6.7.0
celery>=5.4.0
redis>=5.0.8
cryptography>=43.0.0
django-filter>=24.0
drf-spectacular>=0.27.0
